ARM: Stricten arguments to virt_to_pfn()
The kernel convention is that the argument passed to virt_to_pfn()
(virtual address to page frame number) needs to be an unsigned long,
but ARM has avoided this by including a cast to (unsigned long)
inside the definition of virt_to_pfn() itself. This creates problems
if we want to make virt_to_pfn() an actual function with a typed argument,
so fix up the offending call sites.

+++ arch/arm/kernel/hibernate.c
@@ -25,8 +25,8 @@
 
 int pfn_is_nosave(unsigned long pfn)
 {
-	unsigned long nosave_begin_pfn = virt_to_pfn(&__nosave_begin);
-	unsigned long nosave_end_pfn = virt_to_pfn(&__nosave_end - 1);
+	unsigned long nosave_begin_pfn = virt_to_pfn((unsigned long)&__nosave_begin);
+	unsigned long nosave_end_pfn = virt_to_pfn((unsigned long)(&__nosave_end - 1));
 
 	return (pfn >= nosave_begin_pfn) && (pfn <= nosave_end_pfn);
 }
